add missing tests for remove parent status endpoint
