+++ server/routes/admin/remove-parent-status/index.js
@@ -6,13 +6,16 @@ const removeParentStatus = async (req, res) => {
   try {
     const { establishmentId } = req.body;
     const workplace = await models.establishment.findByUid(establishmentId);
+    if (!workplace) {
+      return res.status(400).json({ message: 'Establishment could not be found' });
+    }
     await models.establishment.updateEstablishment(workplace.id, {
       isParent: false,
     });
     res.status(200).send();
   } catch (error) {
     console.error(error);
-    res.status(500).send({ error });
+    res.status(500).json({ message: 'There was an error removing parent status' });
   }
 };
